third_party/WebKit/LayoutTests/http/tests/navigation/pushstate-at-unique-origin-denied.html

<meta http-equiv="Content-Security-Policy" content="sandbox allow-scripts">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function () {
    assert_throws('SecurityError', function () {
        history.pushState(null, null, document.URL);
    });
}, 'pushState at unique origin should fail with SecurityError');
</script>
